supporting long long input in criteo2hugectr.cpp

+++ tools/criteo_script/criteo2hugectr.cpp
@@ -150,7 +150,7 @@ int main(int argc, char* argv[]){
       
       data_writer.append(reinterpret_cast<char*>(&KEYS_WIDE_MODEL), sizeof(int));
       for(int j = dense_dim+label_dim; j < KEYS_WIDE_MODEL+dense_dim+label_dim; j++){
-	T key = static_cast<T>(std::stoi(vec_string[j]));
+	T key = static_cast<T>(std::stoll(vec_string[j]));
 	data_writer.append(reinterpret_cast<char*>(&key), sizeof(T));
 #ifndef NDEBUG
 	std::cout << key << ',';
@@ -158,7 +158,7 @@ int main(int argc, char* argv[]){
       }
       int nnz = 1;
       for(int j = KEYS_WIDE_MODEL+dense_dim+label_dim; j <  KEYS_WIDE_MODEL+KEYS_DENSE_MODEL+dense_dim+label_dim; j++){
-        T key = static_cast<T>(std::stoi(vec_string[j]));
+        T key = static_cast<T>(std::stoll(vec_string[j]));
 	data_writer.append(reinterpret_cast<char*>(&nnz), sizeof(int));
 	data_writer.append(reinterpret_cast<char*>(&key), sizeof(T));
 #ifndef NDEBUG
